requirements.txt
chromadb==1.5.2
datasets==4.6.1
groq==1.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
sentence-transformers==5.2.3
streamlit==1.54.0
tqdm==4.67.3
